renamed tons and tons of functions
