+++ pinger_finder/bbb/LTC1564.py
@@ -1,4 +1,4 @@
-from . import BBBIO
+from .port import Port
 
 CS_PIN = ''
 F_PINS = ''
@@ -10,38 +10,39 @@ DEFAULT_CS = 1
 
 
 class LTC1564():
+    """
+    """
 
     def __init__(self):
         # Init all pins
         if F_PINS:
-            self.F = BBBIO.Port(F_PINS)
+            self.F = Port(F_PINS)
             self.F.setPortDir('out')
             self.F.writeToPort(DEFAULT_F)
 
-        self.G = BBBIO.Port(G_PINS)
+        self.G = Port(G_PINS)
         self.G.setPortDir('out')
         self.G.writeToPort(DEFAULT_G)
 
         if CS_PIN:
-            self._CS = BBBIO.Port(CS_PIN)
+            self._CS = Port(CS_PIN)
             self._CS.setPortDir('out')
             self._CS.writeToPort(DEFAULT_CS)
 
             # Init parameter
         if F_PINS:
             self.Fval = DEFAULT_F
-
             self.Gval = DEFAULT_G
 
-    def GainMode(self, mode):
+    def gain_mode(self, mode):
         """ Configures gain of the input stage.
 
             Args:
                 mode: int
         """
-        n = self.GetNGainStates()
+        n = self.get_n_gain_states()
 
-        if (0 <= mode <= n - 1):  # n-1 is the max bit value that can be used
+        if 0 <= mode <= n - 1:  # n-1 is the max bit value that can be used
             print("LTC1564: Writing %d to gain stage." % mode)
 
             if CS_PIN:
@@ -53,14 +54,15 @@ class LTC1564():
                 self._CS.writeToPort(1)
 
             self.Gval = mode
-
         else:
             print("LTC1564: mode %s is outside the range of possible gain states" % mode)
 
-    def GetNGainStates(self):
-        return 2**(self.G.length)
+    def get_n_gain_states(self):
+        """
+        """
+        return 2 ** self.G.length
 
-    def FiltMode(self, mode):
+    def filter_mode(self, mode):
         """ Configures Fc of the input stage
 
             Args:
@@ -72,7 +74,9 @@ class LTC1564():
 
             if CS_PIN:
                 self._CS.writeToPort(0)
+
             self.F.writeToPort(mode)
+
             if CS_PIN:
                 self._CS.writeToPort(1)
 
@@ -80,7 +84,7 @@ class LTC1564():
         else:
             print("LTC1564: mode %s is outside the range of possible filt states" % mode)
 
-    def Terminate(self):
+    def terminate(self):
         """ Terminates control of the LTC1564.
 
         WARNING: Careless use of this cmd is rather discouraged unless
