+++ pinger_finder/bbb/__init__.py
@@ -1,0 +1,2 @@
+
+BIN_DIR = "pru/bin/"
