+++ pinger_finder/bbb/pru/bin/.gitignore
