+++ pinger_finder/bbb/pru/init0.p
