+++ pinger_finder/bbb/pru/init1.p
