+++ pinger_finder/bbb/pru/pru1.p
